Refactoring: BoardConfiguration was in fack just a checker if specyfic configuration can be used.
